Isolate and condition IAM Role creation with helper
All IAM Roles have been deterministially named (based on stack and
ProjectName) and consolidated in a single iam-roles template. This will
allow a high-privilege account user to pre-deploy the necessary IAM
resources into a more tightly controlled account.

A helper function is included that is able to inspect the account for
the necessary roles and retrieve the ARNs that are then passed to the
relevant nested stacks. When a nested stack is passed an ARN, it skips
the creation of the IAM Role and refers to the referenced resource.

NOTE: There is a chicken and egg problem in that the role_checker
function cannot inspect the account without being assigned a
sufficiently privileged role. We cannot create this role for the
role_checker, nor reference a pre-created role by default as most users
will opt to let the QS create all roles. To work around this, we require
the user to explicilty pass IamInspectRoleArn as a parameter if they
have elected to pre-deploy IAM resources and take advantage of
role_checker.

This is not quite complete and still requires some polish/testing